backend/requirements.txt
django>=5.1.5
djangorestframework>=3.15.2
django-cors-headers>=4.3.1
django-filter>=24.3
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.4.0
recipe-scrapers==15.4.0
